refactor: use feature folders
