+++ server/middlewares/with-authentication.ts
